requirements.txt
# Core ML
scikit-learn>=1.3
xgboost>=2.0
numpy>=1.24
pandas>=1.5

# MLflow
mlflow>=2.9

# API
fastapi>=0.104
uvicorn[standard]>=0.24
pydantic>=2.5

# Monitoring
prometheus-client>=0.19

# Configuration
pyyaml>=6.0
python-dotenv>=1.0

# Dev / Testing
pytest>=7.0
pytest-cov>=4.0
pytest-asyncio>=0.23
httpx>=0.25
ruff>=0.1
pre-commit>=3.0
